Several changes
1. Reformatting

2. Fixing semantics of indices() return values.

+++ lib/coek/coek/api/variable_array.cpp
@@ -124,7 +124,7 @@ Variable VariableArray::index(const IndexVector& args)
     // We know that the args[i] values are nonnegative b.c. we have asserted that while
     // processing these arguments
     size_t ndx = static_cast<size_t>(args[0]);
-    for (size_t i = 1; i < args.size(); i++) ndx = ndx * shape[i] + args[i];
+    for (size_t i = 1; i < args.size(); i++) ndx = ndx * shape[i] + static_cast<size_t>(args[i]);
 
     if (ndx > size()) {
         std::string err = "Unknown index value: " + _repn->variable_template.name() + "[";
